Complete revamp of script
This is a completely new script. Instead of using Selenium, this one relies only on Requests for accessing the web, and BeautifulSoup for parsing the HTML requests retrieves. The payload it generates for getting images also somehow bypasses password prompts, so you can download images that are behind a password without knowing the password... Sorry for the mess, I'm not a developer at all...